+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/preferences/SingletonPreferencesViewModel.kt
@@ -82,7 +82,8 @@ open class SingletonPreferencesViewModel: BaseViewModel<PreferencesContract.Even
             copy(
                 ingredients = reloadedIngredients,
                 diets = diets.map { it.resetWith(dietsOrEmptyFromLocal().map { tag -> tag.id }) },
-                equipments = equipments.map { it.resetWith(equipmentsOrEmptyFromLocal().map { tag -> tag.id }) }
+                equipments = equipments.map { it.resetWith(equipmentsOrEmptyFromLocal().map { tag -> tag.id }) },
+                guests = guestOrNullFromLocal()
             )
         }
     }
@@ -106,6 +107,10 @@ open class SingletonPreferencesViewModel: BaseViewModel<PreferencesContract.Even
 
     private fun equipmentsOrEmptyFromLocal(): List<Tag> {
         return tagsFromLocal(LOCAL_EQUIPMENT_KEY)
+    }
+
+    fun guestOrNullFromLocal(): Int? {
+        return userPreferences.getIntOrNull(LOCAL_GUEST_KEY) 
     }
 
     private fun tagsFromLocal(preferenceKey: String): List<Tag> {
@@ -145,6 +150,7 @@ open class SingletonPreferencesViewModel: BaseViewModel<PreferencesContract.Even
         userPreferences.putList(LOCAL_DIET_KEY, toSaveInStorageSerializedTags(currentState.diets))
         userPreferences.putList(LOCAL_INGREDIENT_KEY, toSaveInStorageSerializedTags(currentState.ingredients))
         userPreferences.putList(LOCAL_EQUIPMENT_KEY, toSaveInStorageSerializedTags(currentState.equipments))
+        currentState.guests?.let { userPreferences.putInt(LOCAL_GUEST_KEY, it) }
         launch(coroutineHandler) { sideEffect.emit(PreferencesEffect.PreferencesChanged) }
     }
 
@@ -163,7 +169,7 @@ open class SingletonPreferencesViewModel: BaseViewModel<PreferencesContract.Even
         return recipeRepositoryImp.getRecipeNumberOfResult(getPreferencesAsQueryString())
     }
 
-    fun changeGlobaleGuest(numberOfGuest: Int) {
+    fun changeGlobalGuest(numberOfGuest: Int) {
         if (numberOfGuest in 1..100) {
             setState { copy(guests = numberOfGuest) }
         }
@@ -184,6 +190,7 @@ open class SingletonPreferencesViewModel: BaseViewModel<PreferencesContract.Even
         val LOCAL_DIET_KEY = "MIAM_${TagTypes.DIET.name}"
         val LOCAL_INGREDIENT_KEY = "MIAM_${TagTypes.INGREDIENT.name}"
         val LOCAL_EQUIPMENT_KEY = "MIAM_${TagTypes.EQUIPMENT.name}"
+        val LOCAL_GUEST_KEY = "MIAM_GUEST"
 
         val defaultIngredientTagIds = listOf(
             "ingredient_category_lgumes",
@@ -201,7 +208,7 @@ open class SingletonPreferencesViewModel: BaseViewModel<PreferencesContract.Even
                 ingredients = emptyList(),
                 equipments = emptyList(),
                 recipesFound = 0,
-                guests = 4
+                guests = null
             )
         }
     }