+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/recipe/RecipeContract.kt
@@ -1,7 +1,9 @@
 package com.miam.kmmMiamCore.component.recipe
 
 import com.miam.kmmMiamCore.base.mvi.*
+import com.miam.kmmMiamCore.handler.GroceriesListHandler.groceriesListStore
 import com.miam.kmmMiamCore.miam_core.model.Recipe
+import com.miam.kmmMiamCore.services.UserPreferencesInstance
 
 
 enum class TabEnum {
@@ -49,13 +51,13 @@ interface RecipeContract {
                 ?: false
         }
 
-        private fun retrieveGuest(isInCart: Boolean, groceriesListStore: GroceriesListStore): Int {
+        private fun retrieveGuest(isInCart: Boolean, groceriesListSingletonPreferencesViewModelStore: GroceriesListStore): Int {
             if (isInCart) {
                 val currentGl = groceriesListStore.observeState().value.groceriesList
                 return (currentGl?.attributes?.recipesInfos?.find { ri -> ri.id.toString() == recipe?.id })?.guests
                     ?: 4
             }
-            return recipe?.attributes?.numberOfGuests ?: 4
+            return UserPreferencesInstance.instance.getIntOrNull("MIAM_GUEST") ?: recipe?.attributes?.numberOfGuests ?: 4
         }
     }