+++ changelog.md
@@ -3,6 +3,7 @@
 
 # 3.4.1
 - [FIX][COR] Report 3.3.1 fix on filter to 3.4
+- [FEA][COR] Router service previous now return new current route
 
 # 3.4.0
 - [FEA][COR] New Route Service